+++ src/main/kotlin/com/github/insanusmokrassar/TelegramBotAPI/requests/edit/LiveLocation/EditInlineMessageLiveLocation.kt
@@ -5,8 +5,8 @@ import com.github.insanusmokrassar.TelegramBotAPI.types.*
 import com.github.insanusmokrassar.TelegramBotAPI.types.buttons.InlineKeyboardMarkup
 import com.github.insanusmokrassar.TelegramBotAPI.types.message.RawMessage
 import kotlinx.serialization.*
+import kotlinx.serialization.internal.BooleanSerializer
 
-// TODO:: Replace return type by RawMessage or Boolean (as in documentation)
 @Serializable
 data class EditInlineMessageLiveLocation(
     @SerialName(inlineMessageIdField)
@@ -20,5 +20,4 @@ data class EditInlineMessageLiveLocation(
     override val replyMarkup: InlineKeyboardMarkup? = null
 ) : EditInlineMessage, EditReplyMessage, EditLocationMessage {
     override fun method(): String = "editMessageLiveLocation"
-    override fun resultSerializer(): KSerializer<RawMessage> = RawMessage.serializer()
 }