+++ src/main/kotlin/com/github/insanusmokrassar/TelegramBotAPI/requests/edit/abstracts/EditTextChatMessage.kt
@@ -1,10 +1,8 @@
 package com.github.insanusmokrassar.TelegramBotAPI.requests.edit.abstracts
 
-import com.github.insanusmokrassar.TelegramBotAPI.requests.abstracts.SimpleRequest
 import com.github.insanusmokrassar.TelegramBotAPI.types.ParseMode.ParseMode
-import com.github.insanusmokrassar.TelegramBotAPI.types.message.RawMessage
 
-interface EditTextChatMessage : SimpleRequest<RawMessage> {
+interface EditTextChatMessage {
     val text: String
     val parseMode: ParseMode?
 }